third_party/WebKit/LayoutTests/fast/forms/setrangetext-within-events-expected.txt

This tests the selection of the text field after setSelectionRange is called

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
focus test :
PASS textfield.selectionStart is 0
PASS textfield.selectionEnd is 5
the selection should be cleared
PASS textfield.selectionStart is 0
PASS textfield.selectionEnd is 0
mousedown test :
PASS textfield.selectionStart is 0
PASS textfield.selectionEnd is 0
the selection should be cleared
PASS textfield.selectionStart is 0
PASS textfield.selectionEnd is 0
mouseup test :
PASS textfield.selectionStart is 0
PASS textfield.selectionEnd is 5
the selection should be cleared
PASS textfield.selectionStart is 0
PASS textfield.selectionEnd is 0
click test :
PASS textfield.selectionStart is 0
PASS textfield.selectionEnd is 5
the selection should be cleared
PASS textfield.selectionStart is 0
PASS textfield.selectionEnd is 0
